requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
click==8.1.3
colorama==0.4.6
psycopg2-binary==2.9.6
tzdata==2023.3
